refactor(core): move utility functions to utils module
index.js:
  - remove opt-out file logic (moved to utils/optOut.js)
  - remove deepEqual function (moved to utils/helpers.js)
  - remove cors origin check logic (moved to utils/helpers.js)
  - remove getMember and formatUserData functions (moved to utils/helpers.js)
  - import all utilities from utils module
utils/optOut.js:
  - add ensureOptOutFile
  - add loadOptOutList, saveOptOutList
  - add isOptedOut, optOutUser, optInUser
utils/helpers.js:
  - add deepEqual function
  - add corsOriginCheck function
  - add getMember and formatUserData functions
  - add internal getActivityTypeName helper
utils/index.js:
  - re-export helpers from optOut.js and helpers.js
